theme and window structure adjustments + scrolling
the window's handle/header bar is now offset up from the containing panel,
changed player stuff to unhandled input so that the UI will catch inputs and handle them so e.g. scrolling on a window doesn't zoom the game in the background.
the window header label now clips with ellipsis
window content now added inside a scroll container so that the window becomes scrollable when smaller than its contents.